Fixed tiny build issue on linux

+++ app/mon/mon_gui/src/widgets/models/service_tree_item.cpp
@@ -111,7 +111,7 @@ QVariant ServiceTreeItem::data(Columns column, Qt::ItemDataRole role) const
       }
       else if (column == Columns::CALL_COUNT)
       {
-        return method_.call_count();
+        return (long long)method_.call_count();
       }
       else
       {
